requirements.txt
flask
crewai
groq
sqlalchemy
langchain_community
requests
